% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indipsa.R
\docType{data}
\name{indipsa}
\alias{indipsa}
\title{Selective Stock Price Index}
\format{
A time series object with 3186 elements
\describe{It's a stock market index that tracks the performance of a select group of Chilean companies.}
}
\source{
Yahoo Finance
}
\usage{
indipsa
}
\description{
The data covers the period from March 2000 to October 2014, totaling 3186 observations.
}
\keyword{datasets}
