% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGARCH_Sim.R
\name{LSGARCH_Sim}
\alias{LSGARCH_Sim}
\title{Series simulation of a LSGARCH model.}
\usage{
LSGARCH_Sim(
  n,
  gamma,
  alpha,
  beta,
  type = c("polynomial", "NoLineal", "trigonometric"),
  exponentes = NULL,
  trig = NULL,
  arg = NULL
)
}
\arguments{
\item{n}{Long to simulate the series}

\item{gamma}{Vector containing coefficents of c.}

\item{alpha}{Vector containing coefficents of alpha.}

\item{beta}{Vector containing coefficents of beta.}

\item{type}{Vector of function type for c, alpha and beta.}

\item{exponentes}{Vector for exponenets in NoLineal.}

\item{trig}{Type of trigonometric function.}

\item{arg}{Value of argument for the trigonometric function.}
}
\value{
Vector with two componentes, where the first component is dependent variable and the second component is independent variable.
}
\description{
Function to simulate a series of a long n LSGARCH model.
}
