% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSgarchKalmanFit.R
\name{LSgarchKalmanFit}
\alias{LSgarchKalmanFit}
\title{Fit LSGARCH parameters with Kalman Filter.}
\usage{
LSgarchKalmanFit(
  series,
  c,
  alpha,
  beta,
  type = c("polynomial", "NoLineal", "trigonometric"),
  exponentes,
  trig,
  arg,
  predict = 0,
  trace.log = F
)
}
\arguments{
\item{series}{Time series.}

\item{c}{Vector containing coefficents of c.}

\item{alpha}{Vector containing coefficents of alpha.}

\item{beta}{Vector containing coefficents of beta.}

\item{type}{Vector of function type for c, alpha and beta.}

\item{exponentes}{Vector for exponenets in NoLineal.}

\item{trig}{Type of trigonometric function.}

\item{arg}{Value of argument for the trigonometric function.}

\item{predict}{Value for time to generate predict.}

\item{trace.log}{Variable to print names of coefficients.}
}
\value{
Return fit values of omega, alpha and beta
}
\description{
Fit procedure for coefficents of c, \eqn{\alpha} and \eqn{\beta}.
}
