% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSgarchKalmanPrint.R
\name{LSgarchKalmanPrint}
\alias{LSgarchKalmanPrint}
\title{LS-TvGarch Filter Kalman print outputs.}
\usage{
LSgarchKalmanPrint(
  x,
  series,
  c,
  alpha,
  beta,
  nsample = length(series),
  type = c("polynomial", "NoLineal", "trigonometric"),
  exponentes,
  trig,
  arg,
  trace.log = F,
  predict
)
}
\arguments{
\item{x}{Vector of coefficents to fit.}

\item{series}{Time series.}

\item{c}{Vector containing coefficents of c.}

\item{alpha}{Vector containing coefficents of alpha.}

\item{beta}{Vector containing coefficents of beta.}

\item{nsample}{Value of time series length.}

\item{type}{Vector of function type for c, alpha and beta.}

\item{exponentes}{Vector for exponenets in NoLineal.}

\item{trig}{Type of trigonometric function.}

\item{arg}{Value of argument for the trigonometric function.}

\item{trace.log}{Variable to print names of coefficients.}

\item{predict}{Value for time to generate predict.}
}
\value{
A data frame containing the following columns:
\itemize{
\item \code{X}: State vector of Kalman equations.
\item \code{Fm}: Value of MSE
\item \code{sigma}: Conditional variance.
\item \code{loglike}: Value of the loglike.
}
}
\description{
That function it is for print outputs of tvGarch model
who contains the returns, variance conditional, value of log likelihood, and MSE.
}
