% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvGarchKalmanLoglike.R
\name{tvGarchKalmanLoglike}
\alias{tvGarchKalmanLoglike}
\title{Models tv-Garch Filter Kalman LogLikehood.}
\usage{
tvGarchKalmanLoglike(
  x,
  series,
  c,
  alpha,
  beta,
  nsample = length(series),
  type = c("polynomial", "NoLineal", "trigonometric"),
  exponentes,
  trig,
  arg,
  predict,
  trace.log = F
)
}
\arguments{
\item{x}{Vector of coefficents to fit.}

\item{series}{Time series.}

\item{c}{Vector containing coefficents of c.}

\item{alpha}{Vector containing coefficents of alpha.}

\item{beta}{Vector containing coefficents of beta.}

\item{nsample}{Value of time series length.}

\item{type}{Vector of function type for c, alpha and beta.}

\item{exponentes}{Vector for exponenets in NoLineal.}

\item{trig}{Type of trigonometric function.}

\item{arg}{Value of argument for the trigonometric function.}

\item{predict}{Value for time to generate predict.}

\item{trace.log}{Variable to print names of coefficients.}
}
\value{
Value of loglike in model.
}
\description{
It is the function to use in the process to fit coefficients.
}
